+++ test/cookie_jar_test.dart
@@ -20,28 +20,76 @@ void main() async {
 
   final dir = Directory('./cookies');
   await dir.create();
+
   group('read and save', () {
     test('DefaultCookieJar', () async {
       final cj = CookieJar();
-      await cj.saveFromResponse(Uri.parse('https://www.baidu.com/xx'), cookies);
-      var results =
-          await cj.loadForRequest(Uri.parse('https://www.baidu.com/xx'));
+      var domain = 'https://aa.com';
+      await cj.saveFromResponse(Uri.parse('$domain/a'), cookies);
+      var results = await cj.loadForRequest(Uri.parse('$domain/a'));
       expect(results.length, 2);
-      results =
-          await cj.loadForRequest(Uri.parse('https://www.baidu.com/xx/dd'));
+      results = await cj.loadForRequest(Uri.parse('$domain/a/b'));
       expect(results.length, 2);
-      results = await cj.loadForRequest(Uri.parse('https://www.baidu.com/'));
+      results = await cj.loadForRequest(Uri.parse('$domain/'));
+      expect(results.length, 2);
+      results = await cj.loadForRequest(Uri.parse(domain));
+      expect(results.length, 2);
+
+      domain = 'https://bb.com';
+      await cj.saveFromResponse(Uri.parse('$domain/a/b'), cookies);
+      results = await cj.loadForRequest(Uri.parse('$domain/a'));
+      expect(results.length, 2);
+      results = await cj.loadForRequest(Uri.parse('$domain/a/'));
+      expect(results.length, 2);
+      results = await cj.loadForRequest(Uri.parse('$domain/a/b'));
+      expect(results.length, 2);
+      results = await cj.loadForRequest(Uri.parse('$domain/a/b/c'));
+      expect(results.length, 2);
+      results = await cj.loadForRequest(Uri.parse('$domain/c'));
       expect(results.length, 0);
-      await cj.saveFromResponse(
-          Uri.parse('https://google.com'), cookiesExpired);
-      results = await cj.loadForRequest(Uri.parse('https://google.com'));
+
+      //same key exists in different path
+      domain = 'https://cc.com';
+      await cj.saveFromResponse(Uri.parse('$domain'), [Cookie('a', '1')]);
+      await cj.saveFromResponse(Uri.parse('$domain/a/b'), [Cookie('a', '2')]);
+      results = await cj.loadForRequest(Uri.parse('$domain'));
+      expect(results.length, 1);
+      expect(results.first.value, '1');
+      results = await cj.loadForRequest(Uri.parse('$domain/a/b'));
+      expect(results.length, 2);
+      expect(results.first.value, '2');
+      expect(results.last.value, '1');
+      results = await cj.loadForRequest(Uri.parse('$domain/a/b/c'));
+      expect(results.length, 2);
+      expect(results.first.value, '2');
+      expect(results.last.value, '1');
+
+      domain = 'https://dd.com';
+      await cj.saveFromResponse(Uri.parse(domain), cookiesExpired);
+      results = await cj.loadForRequest(Uri.parse(domain));
       expect(results.length, 2);
       await Future<void>.delayed(const Duration(seconds: 2), () async {
-        results = await cj.loadForRequest(Uri.parse('https://google.com'));
+        results = await cj.loadForRequest(Uri.parse(domain));
         expect(results.length, 1);
       });
     });
+    test('PersistCookieJar', () async {
+      var domain = 'https://aa.com';
+      // write
+      final pcj = PersistCookieJar(storage: FileStorage('./test/cookies'));
+      await pcj.saveFromResponse(Uri.parse('$domain/a'), cookies);
 
+      // read
+      final cj = PersistCookieJar(storage: FileStorage('./test/cookies'));
+      var results = await cj.loadForRequest(Uri.parse('$domain/a'));
+      expect(results.length, 2);
+      results = await cj.loadForRequest(Uri.parse('$domain/a/b'));
+      expect(results.length, 2);
+      results = await cj.loadForRequest(Uri.parse('$domain/'));
+      expect(results.length, 2);
+      results = await cj.loadForRequest(Uri.parse(domain));
+      expect(results.length, 2);
+    });
     test('SharedCookie', () async {
       final cj = CookieJar();
       final cookies = <Cookie>[
@@ -55,31 +103,37 @@ void main() async {
       expect(results.length, 1);
     });
 
-    test('SharedPathCookie', () async {
-      final cj = CookieJar();
-      final cookies = <Cookie>[
-        Cookie('JSESSIONID', 'wendux')..path = '/docs',
-      ];
-      await cj.saveFromResponse(Uri.parse('http://www.mozilla.org/'), cookies);
-
-      final results1 = await cj.loadForRequest(Uri.parse('http://www.mozilla.org/docs'));
-      expect(results1.length, 1);
-
-      final results2 = await cj.loadForRequest(Uri.parse('http://www.mozilla.org/docs/'));
-      expect(results2.length, 1);
-
-      final results3 = await cj.loadForRequest(Uri.parse('http://www.mozilla.org/docs/Web'));
-      expect(results3.length, 1);
-
-      final results4 = await cj.loadForRequest(Uri.parse('http://www.mozilla.org/docs/Web/HTTP'));
-      expect(results4.length, 1);
-
-      final results5 = await cj.loadForRequest(Uri.parse('http://www.mozilla.org/docsets'));
-      expect(results5.length, 0);
-
-      final results6 = await cj.loadForRequest(Uri.parse('http://www.mozilla.org/fr/docs'));
-      expect(results6.length, 0);
-    });
+    // test('SharedPathCookie', () async {
+    //   final cj = CookieJar();
+    //   final cookies = <Cookie>[
+    //     Cookie('JSESSIONID', 'wendux')..path = '/docs',
+    //   ];
+    //   await cj.saveFromResponse(Uri.parse('http://www.mozilla.org/'), cookies);
+    //
+    //   final results1 =
+    //       await cj.loadForRequest(Uri.parse('http://www.mozilla.org/docs'));
+    //   expect(results1.length, 1);
+    //
+    //   final results2 =
+    //       await cj.loadForRequest(Uri.parse('http://www.mozilla.org/docs/'));
+    //   expect(results2.length, 1);
+    //
+    //   final results3 =
+    //       await cj.loadForRequest(Uri.parse('http://www.mozilla.org/docs/Web'));
+    //   expect(results3.length, 1);
+    //
+    //   final results4 = await cj
+    //       .loadForRequest(Uri.parse('http://www.mozilla.org/docs/Web/HTTP'));
+    //   expect(results4.length, 1);
+    //
+    //   final results5 =
+    //       await cj.loadForRequest(Uri.parse('http://www.mozilla.org/docsets'));
+    //   expect(results5.length, 0);
+    //
+    //   final results6 =
+    //       await cj.loadForRequest(Uri.parse('http://www.mozilla.org/fr/docs'));
+    //   expect(results6.length, 0);
+    // });
 
     test('SharedDomainCookie', () async {
       final cj = CookieJar();
@@ -88,16 +142,20 @@ void main() async {
       ];
       await cj.saveFromResponse(Uri.parse('http://www.mozilla.org/'), cookies);
 
-      final results1 = await cj.loadForRequest(Uri.parse('http://mozilla.org/'));
+      final results1 =
+          await cj.loadForRequest(Uri.parse('http://mozilla.org/'));
       expect(results1.length, 1);
 
-      final results2 = await cj.loadForRequest(Uri.parse('http://developer.mozilla.org/'));
+      final results2 =
+          await cj.loadForRequest(Uri.parse('http://developer.mozilla.org/'));
       expect(results2.length, 1);
 
-      final results3 = await cj.loadForRequest(Uri.parse('http://fakemozilla.org/'));
+      final results3 =
+          await cj.loadForRequest(Uri.parse('http://fakemozilla.org/'));
       expect(results3.length, 0);
 
-      final results4 = await cj.loadForRequest(Uri.parse('http://mozilla.org.com/'));
+      final results4 =
+          await cj.loadForRequest(Uri.parse('http://mozilla.org.com/'));
       expect(results4.length, 0);
     });
 
@@ -109,11 +167,13 @@ void main() async {
       await cj.saveFromResponse(Uri.parse('http://www.mozilla.org/'), cookies);
 
       await cj.delete(Uri.parse('http://www.fakemozilla.org/'), true);
-      final results1 = await cj.loadForRequest(Uri.parse('http://www.mozilla.org/'));
+      final results1 =
+          await cj.loadForRequest(Uri.parse('http://www.mozilla.org/'));
       expect(results1.length, 1);
 
       await cj.delete(Uri.parse('http://developer.mozilla.org/'), true);
-      final results2 = await cj.loadForRequest(Uri.parse('http://www.mozilla.org/'));
+      final results2 =
+          await cj.loadForRequest(Uri.parse('http://www.mozilla.org/'));
       expect(results2.length, 0);
     });
 
@@ -132,56 +192,16 @@ void main() async {
       results =
           await cj.loadForRequest(Uri.parse('https://tt.facebook.com/xxx'));
       expect(results.length, 1);
-//      await cj.delete(Uri.parse('https://www.facebook.com/'), true);
-//      results =
-//          await cj.loadForRequest(Uri.parse('https://tt.facebook.com/xxx'));
-//      expect(results.length, 0);
-    });
-
-    test('PersistCookieJar', () async {
-      final cj = PersistCookieJar(storage: FileStorage('./test/cookies'));
-      await cj.saveFromResponse(Uri.parse('https://www.baidu.com/xx'), cookies);
-      var results =
-          await await cj.loadForRequest(Uri.parse('https://www.baidu.com/xx'));
-      expect(results.length, 2);
-      results =
-          await cj.loadForRequest(Uri.parse('https://www.baidu.com/xx/dd'));
-      expect(results.length, 2);
-      results = await cj.loadForRequest(Uri.parse('https://www.baidu.com/'));
-      expect(results.length, 0);
-      await cj.delete(Uri.parse('https://www.baidu.com/'));
-      results =
-          await cj.loadForRequest(Uri.parse('https://www.baidu.com/xx/dd'));
-      expect(results.length, 0);
-      await cj.saveFromResponse(Uri.parse('https://www.baidu.com/xx'), cookies);
-      await cj.saveFromResponse(
-          Uri.parse('https://google.com'), cookiesExpired);
-      results = await cj.loadForRequest(Uri.parse('https://google.com'));
-      expect(results.length, 2);
-      await Future<void>.delayed(const Duration(seconds: 2), () async {
-        results = await cj.loadForRequest(Uri.parse('https://google.com'));
-        expect(results.length, 1);
-      });
-    });
-
-    test('PersistCookieJarLoad', () async {
-      final cj = PersistCookieJar(storage: FileStorage('./test/cookies'));
-      var results =
-          await cj.loadForRequest(Uri.parse('https://www.baidu.com/xx'));
-      expect(results.length, 2);
-      results =
-          await cj.loadForRequest(Uri.parse('https://www.baidu.com/xx/dd'));
-      expect(results.length, 2);
-      results = await cj.loadForRequest(Uri.parse('https://www.baidu.com/'));
-      expect(results.length, 0);
-      results = await cj.loadForRequest(Uri.parse('https://google.com'));
     });
 
     test('PersistCookieIgnoreExpires', () async {
       var cj = PersistCookieJar(
-          ignoreExpires: true, storage: FileStorage('./test/cookies'));
-      final uri = Uri.parse('https://xxx.xxx.com/');
+        ignoreExpires: true,
+        storage: FileStorage('./test/cookies'),
+      );
+      final uri = Uri.parse('https://xxx.com/');
       await cj.delete(uri);
+
       List<Cookie> results;
       final cookie = Cookie('test', 'hh')
         ..expires = DateTime.parse('1970-02-27 13:27:00');
@@ -191,7 +211,9 @@ void main() async {
       results = await cj.loadForRequest(uri);
       expect(results.length, 1);
       cj = PersistCookieJar(
-          ignoreExpires: false, storage: FileStorage('./test/cookies'));
+        ignoreExpires: false,
+        storage: FileStorage('./test/cookies'),
+      );
       results = await cj.loadForRequest(uri);
       expect(results.length, 0);
     });