+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 4.0.1-beta
+- Preserve cookies that with same name but in different paths when request (as Chrome).
+  eg(in request header): Cookie: a=1;  a=2;  a=3
+- Fix path and domain match error (#30)
+- Fix#31: Use the "directory" of request-uri as default path if there is no 'Path' component in Set-Cookie line (in 3.x, used root '/', it is not standard).
 ## 3.0.1
 
 - refactor